Fixed sql insertions, variable names, other suggested changes
